Introduce Node Killing as a spec type
This builds on the prior refactoring to add the ability to
kill nodes as a "type" of chaos specification. It works,
simply, by asking for the node to be deleted.

This has been tested in GKE and, with automatic node pool
management shenanigans, whatever they are called - I'm flying
so I can't google terms so I know nothing - this brings the
nodes down, pods migrate, and GKE brings new nodes back.

The monkey can only do one thing at a time still, so to have
it, say, both kill nodes and kill pods, you need two monkeys.